fix inverse enum mapping

+++ tools/nodejs/src/duckdb_node.cpp
@@ -20,14 +20,14 @@ NodeDuckDB::NodeDuckDB(Napi::Env env, Napi::Object exports) {
 	token_type_enum.Set("OPERATOR", 3);
 	token_type_enum.Set("KEYWORD", 4);
 	token_type_enum.Set("COMMENT", 5);
-	
+
 	// TypeScript enums expose an inverse mapping.
-	token_type_enum.Set(0, "IDENTIFIER");
-	token_type_enum.Set(1, "NUMERIC_CONSTANT");
-	token_type_enum.Set(2, "STRING_CONSTANT");
-	token_type_enum.Set(3, "OPERATOR");
-	token_type_enum.Set(4, "KEYWORD");
-	token_type_enum.Set(5, "COMMENT");
+	token_type_enum.Set((uint32_t)0, "IDENTIFIER");
+	token_type_enum.Set((uint32_t)1, "NUMERIC_CONSTANT");
+	token_type_enum.Set((uint32_t)2, "STRING_CONSTANT");
+	token_type_enum.Set((uint32_t)3, "OPERATOR");
+	token_type_enum.Set((uint32_t)4, "KEYWORD");
+	token_type_enum.Set((uint32_t)5, "COMMENT");
 
 	token_type_enum_ref = Napi::ObjectReference::New(token_type_enum);
